+++ packages/bluesky/lib/bluesky.dart
@@ -148,5 +148,6 @@ export 'package:bluesky/src/params/list_param.dart';
 export 'package:bluesky/src/params/post_param.dart';
 export 'package:bluesky/src/params/repo_param.dart';
 export 'package:bluesky/src/params/strong_ref_param.dart';
+export 'package:bluesky/src/params/thread_param.dart';
 export 'package:bluesky/src/servers/servers_service.dart' show createSession;
 export 'package:bluesky/src/unspecced/unspecced_service.dart';